pyproject.toml
[project]
name = "scripting-pdf"
version = "0.1.0"
description = "AI-powered PDF translation SaaS platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "psycopg2-binary>=2.9.7",
    "PyMuPDF>=1.23.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "pandas>=2.0.0",
    "reportlab>=4.0.0",
    "arabic-reshaper>=3.0.0",
    "python-bidi>=0.4.2",
    "crewai>=0.28.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.0.20",
    "chromadb>=0.4.0",
    "tiktoken>=0.5.0",
    "nltk>=3.8.0",
]
